Fix(listview): Prevent idle crash by correcting ListView dirty-checking logic
This commit resolves a critical stability bug that caused the application to crash after idling in the library view with a large number of games.

The root cause was a flawed dirty-checking mechanism in ⁠CB_ListView_draw. The ⁠CB_ListViewModel, which holds a snapshot of the previous frame's state, used ⁠int for its ⁠scrollIndicatorOffset and ⁠scrollIndicatorHeight fields. However, the live state in ⁠CB_ListViewScroll used ⁠float.

Due to uninitialized memory, the model's integer fields contained garbage values at startup. The comparison between the model's ⁠garbage int and the live state's ⁠valid float would always evaluate to ⁠true, forcing the entire view to be redrawn on every frame. This constant redrawing created by a large game list led to system instability and the eventual crash.

This fix refactors the ⁠ListView component to use a consistent ⁠int data type for all scrollbar pixel coordinates and dimensions (⁠indicatorOffset and ⁠indicatorHeight).
By enforcing type consistency, the dirty-checking comparison becomes a robust ⁠int-vs-int operation. After the first frame, the model's state correctly mirrors the live state, the comparison evaluates to ⁠false, the unnecessary redraws cease, and the application remains stable.

+++ src/listview.h
@@ -32,8 +32,8 @@ typedef struct
     float time;
     float duration;
     bool indicatorVisible;
-    float indicatorOffset;
-    float indicatorHeight;
+    int indicatorOffset;
+    int indicatorHeight;
 } CB_ListViewScroll;
 
 typedef enum
